Added On Conflict in postgresql

+++ routers/Leetcode_Contest/contest_scrape.py
@@ -8,7 +8,7 @@ from sqlalchemy.orm import sessionmaker
 from config import settings
 import time
 from sqlalchemy.exc import IntegrityError
-# from sqlalchemy.dialects.postgresql import insert
+from sqlalchemy.dialects.postgresql import insert
 
 
 DATABASE_URL = SQLALCHEMY_DATABASE_URL = f"postgresql://{settings.DATABASE_USERNAME}:{settings.DATABASE_PASSWORD}@{settings.DATABASE_HOSTNAME}:{settings.DATABASE_PORT}/{settings.DATABASE_NAME}"
@@ -29,25 +29,32 @@ def scrape_each_page(page_number: int,  contest_name: str, db):
             contestants = data["total_rank"]
             submissions = data["submissions"]
             question_ids = [question['question_id'] for question in questions]
+
+            # entries = []
+
             for contestant, submission in zip(contestants, submissions):
                 result = make_dict(contestant, submission, question_ids)
                 contestant_info = models.Contest(**result)
+                # entries.append(contestant_info)
 
-                try:
-                    db.add(contestant_info)
-                    # db.commit()
-                except IntegrityError:
-                    # If there is a conflict (duplicate primary key), update the existing row
-                    db.rollback()  # Rollback the previous transaction
-                    # Merge the new data with the existing row
-                    db.merge(contestant_info)
-                    # db.commit()  # Commit the merged data to the database
+                # try:
+                #     # if len(entries) >= 50:  # Example batch size
+                #     db.add(contestant_info)
+                #     db.commit()
+                #     # entries.clear()
+                # except IntegrityError:
+                #     db.rollback()
+                #     db.merge(contestant_info)
+                #     db.commit()
+                insert_stmt = insert(
+                    models.Contest.__table__).values(**result)
+                do_update_stmt = insert_stmt.on_conflict_do_update(
+                    index_elements=['username'], set_=result)
+                db.execute(do_update_stmt)
 
-                # insert_stmt = insert(
-                #     models.Contest.__table__).values(**result)
-                # do_update_stmt = insert_stmt.on_conflict_do_update(
-                #     index_elements=['username'], set_=result)
-                # db.execute(do_update_stmt)
+            # if entries:
+            #     db.add_all(entries)
+            #     db.commit()
 
             print(f"Page {page_number} done")
     except Exception as e:
@@ -57,25 +64,45 @@ def scrape_each_page(page_number: int,  contest_name: str, db):
 def contest_scrape(contest_name: str):
     print(
         f"https://leetcode.com/contest/api/ranking/{contest_name}/?pagination=1&region=global")
-    total_pages = math.ceil(requests.get(
-        f"https://leetcode.com/contest/api/ranking/{contest_name}/?pagination=1&region=global", proxies=settings.PROXIES, headers=settings.LEETCODE_HEADER).json()["user_num"]/25)
+    response = requests.get(
+        f"https://leetcode.com/contest/api/ranking/{contest_name}/?pagination=1&region=global", proxies=settings.PROXIES, headers=settings.LEETCODE_HEADER)
+    if response.status_code != 200:
+        print(f"Error: {response.status_code}")
+        contest_scrape(contest_name)
+    else:
+        total_pages = math.ceil(response.json()["user_num"]/25)
 
-    print(total_pages)
-    start_time = time.perf_counter()
-    db = SessionLocal()
-    if db.query(models.Contest).count() > 0:
-        db.query(models.Contest).delete()
+        print(total_pages)
+        start_time = time.perf_counter()
+        db = SessionLocal()
+
+        try:
+            # Perform your database operation here
+            if db.query(models.Contest).count() > 0:
+                db.query(models.Contest).delete()
+                db.commit()
+        except Exception as e:
+            # If an exception occurs, rollback the session
+            db.rollback()
+            db.query(models.Contest).delete()
+            print(f"Error occurred: {e}")
+
+        with concurrent.futures.ThreadPoolExecutor() as executor:
+            futures = [executor.submit(
+                scrape_each_page, page, contest_name, db) for page in range(1, total_pages+1)]
+
+        for future in concurrent.futures.as_completed(futures):
+            try:
+                future.result()  # Will raise an exception if the thread failed
+            except Exception as e:
+                print(f"Error in thread: {e}")
+                db.rollback()  # Rollback the entire transaction if any thread fails
+
         db.commit()
 
-    with concurrent.futures.ThreadPoolExecutor() as executor:
-        futures = [executor.submit(
-            scrape_each_page, page, contest_name, db) for page in range(1, total_pages+1)]
-
-    db.commit()
-
-    finish_time = time.perf_counter()
-    print(
-        f"All threads stopped. Finished in {round(finish_time-start_time, 2)} seconds")
+        finish_time = time.perf_counter()
+        print(
+            f"All threads stopped. Finished in {round(finish_time-start_time, 2)} seconds")
 
 
 # contest_scrape("biweekly-contest-126")
